scripts/ingest/requirements.txt
PyYAML==6.0.2
feedparser==6.0.11
python-dateutil==2.9.0.post0
ics==0.7.2
beautifulsoup4==4.12.3
requests==2.32.3
